requirements.txt
fastapi
uvicorn
python-dotenv
pybreaker
requests
httpx
slowapi
pydantic
loadenv